Moves with-toc into plugin file
